Fixed synth output
There was a nasty bug due to the Synth component skipping diff updates
because of the app's tick.
Many diffs could be derived quickly from the played notes, while the
synth would pick only the last one at the end of the burst.
Diffs are now computed in the component so they are in sync with the
tick (which makes more sense anyways as diffs are only used in there).